requirements.txt

chardet==5.2.0
torch==2.3.0
langchain_text_splitters==0.2.2
sentence-transformers==3.0.1
pandas==2.2.2
tqdm==4.66.4
openpyxl==3.1.5
langchain-openai==0.1.14
langchain-core==0.2.33
langchain==0.2.14
python-dotenv==1.0.1
tiktoken==0.7.0
langchain_experimental==0.0.64
tables==3.9.2
docx2txt
tqdm